Desenvolvimento da classe CadastroCliente

+++ src/br/gov/caixa/clientes/ClientePF.java
@@ -4,7 +4,7 @@ import br.gov.caixa.enums.Status;
 
 public final class ClientePF extends Cliente {
 
-    public ClientePF(int id, String nome, Status status) {
+    public ClientePF(long id, String nome, Status status) {
         super(id, nome, status);
     }
 }
