+++ src/br/gov/caixa/contas/ContaCorrente.java
@@ -2,7 +2,7 @@ package br.gov.caixa.contas;
 
 public final class ContaCorrente extends Conta {
 
-    public ContaCorrente(int idUsuario) {
+    public ContaCorrente(long idUsuario) {
         super(idUsuario);
     }
