+++ src/br/gov/caixa/contas/ContaInvestimento.java
@@ -2,7 +2,7 @@ package br.gov.caixa.contas;
 
 public final class ContaInvestimento extends Conta{
 
-    public ContaInvestimento(int idUsuario) {
+    public ContaInvestimento(long idUsuario) {
         super(idUsuario);
     }
