+++ src/br/gov/caixa/contas/ContaPoupanca.java
@@ -1,7 +1,7 @@
 package br.gov.caixa.contas;
 
 public final class ContaPoupanca extends Conta {
-    public ContaPoupanca(int idUsuario) {
+    public ContaPoupanca(long idUsuario) {
         super(idUsuario);
     }
